refactor whitespace handling in the cert/key
